Internal iterators moving to separate directory is in progress.
